# apps/credential-manager/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/credentials/{credentialRecordId}": {
      "get": {
        "operationId": "CredentialsController_get",
        "summary": "Fetch a credential",
        "description": "This call provides a credential for a given tenant",
        "parameters": [
          {
            "name": "credentialRecordId",
            "required": true,
            "in": "path",
            "description": "The credential record ID to retrieve",
            "schema": { "format": "string", "type": "string" }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential fetched successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential fetched successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential fetched successfully",
                      "data": { "id": "71b784a3" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["Credentials"]
      }
    },
...
    "/v1/credential-requests": {
      "get": {
        "operationId": "CredentialRequestsController_find",
        "summary": "Fetch a list of credential requests",
        "description": "This call provides a list of credential requests for a given tenant",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential requests fetched successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential requests fetched successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential requests fetched successfully",
                      "data": [{ "id": "71b784a3" }]
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential request not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential request not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Credential Requests"]
      }
    },
    "/v1/credential-requests/{id}": {
      "get": {
        "operationId": "CredentialRequestsController_getById",
        "summary": "Fetch a credential request by id",
        "description": "This call provides a credential request for a given tenant by id",
        "parameters": [
          {
            "name": "credentialRequestId",
            "required": true,
            "in": "path",
            "description": "The credential request ID to retrieve",
            "schema": { "format": "string", "type": "string" }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential request fetched successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential request fetched successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential request fetched successfully",
                      "data": { "id": "71b784a3" }
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential request not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential request not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Credential Requests"]
      }
    },
...
  },
  "info": {
    "title": "Gaia-X OCM Credential Manager",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}